clean: exe_cmdディレクトリの追加
