+++ srcs/executor/exe_cmd/Makefile
@@ -5,24 +5,39 @@
 #                                                     +:+ +:+         +:+      #
 #    By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2022/03/28 14:21:38 by ynakashi          #+#    #+#              #
-#    Updated: 2022/03/29 15:47:24 by ynakashi         ###   ########.fr        #
+#    Created: 2022/03/29 15:34:53 by ynakashi          #+#    #+#              #
+#    Updated: 2022/03/29 15:53:37 by ynakashi         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
-NAME	= libsignal.a
 
-CC		= cc
+NAME	=	libexe_cmd.a
 
-CFLAGS	= -Wall -Wextra -Werror
+CC		=	cc
 
-RM		= rm -rf
+CFLAGS	=	-Wall -Wextra -Werror
 
-SRCS	= signal.c
+RM		=	rm -rf
 
-OBJS	= $(SRCS:.c=.o)
+SRCS	=	cd.c\
+			no_builtin.c\
+			echo.c\
+			env.c\
+			envlist.c\
+			exit_utils.c\
+			exit.c\
+			export.c\
+			pwd.c\
+			unset.c\
+			utils.c\
+			only_one_cmd.c
 
-.c.o	:
+OBJDIR	:=	./objs
+OBJS := $(addprefix $(OBJDIR)/, $(SRCS:.c=.o))
+
+# suffix rule
+$(OBJDIR)/%.o:	%.c
+	mkdir -p $(OBJDIR)
 	$(CC) $(CFLAGS) -c $< -o $@
 
 $(NAME)	: $(OBJS)
@@ -31,7 +46,9 @@ $(NAME)	: $(OBJS)
 all		: $(NAME)
 
 clean	:
-	$(RM) $(OBJS)
+	if [ -e $(OBJDIR) ]; then \
+		rm -rf $(OBJDIR);\
+	fi
 
 fclean	: clean
 	$(RM) $(NAME)
