+++ srcs/executor/exe_cmd/env.c
@@ -1,28 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pwd.c                                              :+:      :+:    :+:   */
+/*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/25 20:02:03 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/03/25 20:02:07 by ynakashi         ###   ########.fr       */
+/*   Created: 2022/03/25 20:01:02 by ynakashi          #+#    #+#             */
+/*   Updated: 2022/03/29 15:40:59 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "exe_cmd.h"
 
-int	my_pwd(void)
+int	my_env(t_envlist *envlist)
 {
-	char	*pwd_buf;
-
-	pwd_buf = getcwd(NULL, 0);
-	if (pwd_buf == NULL)
+	while (envlist)
 	{
-		print_error("pwd", NULL, errno);
-		return (EXIT_FAILURE);
+		if (envlist->value == NULL)
+		{
+			envlist = envlist->next;
+			continue ;
+		}
+		printf("%s=%s\n", envlist->key, envlist->value);
+		envlist = envlist->next;
 	}
-	ft_putendl_fd(pwd_buf, STDOUT_FILENO);
-	free(pwd_buf);
-	return (EXIT_SUCCESS);
+	return (0);
 }
