+++ srcs/executor/exe_cmd/exe_cmd.h
@@ -1,0 +1,77 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exe_cmd.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/29 15:34:57 by ynakashi          #+#    #+#             */
+/*   Updated: 2022/03/29 15:43:50 by ynakashi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef EXE_CMD_H
+# define EXE_CMD_H
+
+# include "../../../libft/libft.h"
+
+# include <stdio.h> // printf, perror
+# include <stdlib.h> // malloc, free, exit, getenv
+# include <unistd.h> // write, access, read, close, fork, getcwd, chdir, unlink, execve, dup, dup2, pipe, isatty, ttyname, ttyslot
+# include <sys/wait.h> // wait, waitpid, wait3, wait4
+# include <string.h> // strerror
+
+# include <limits.h> // macro
+# include <stdint.h> // macro(linux)
+# include <stdbool.h> // bool
+# include <errno.h> // errno
+
+typedef struct s_envlist
+{
+	char				*key;
+	char				*value;
+	struct s_envlist	*next;
+}	t_envlist;
+
+// envlist.c
+void		free_list(t_envlist *list);
+void		ms_lstadd_back(t_envlist **lst, t_envlist *element);
+t_envlist	*create_envlist(char **envp);
+
+// cd.c
+int		my_cd(char **split_ln, t_envlist **envlist);
+
+// echo.c
+int		my_echo(char **split_ln);
+
+// env.c
+int		my_env(t_envlist *envlist);
+
+// exit.c
+int		my_exit(char **split_ln);
+
+// exit_utils.c
+void	numeric_argument_required(char *str);
+
+// export.c
+void	remove_duplicate(char *str, t_envlist **envlist);
+int		my_export(char **split_ln, t_envlist **envlist);
+
+// pwd.c
+int		my_pwd(void);
+
+// unset.c
+int		my_unset(char **split_ln, t_envlist **envlist);
+
+// no_builtin.c
+int	exe_no_builtin_cmd(char **str, t_envlist *envlist);
+
+// utils.c
+void	safe_free(char **ptr);
+void	free_split(char **ptr);
+void	print_error(char *cmd, char *cmd_arg, int error_number);
+
+// only_one_cmd.c
+int	only_one_cmd(char **dbl_arr, t_envlist **envlist);
+
+#endif // EXE_CMD_H
