+++ srcs/executor/exe_cmd/exit.c
@@ -6,11 +6,11 @@
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 20:01:41 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/03/25 20:01:46 by ynakashi         ###   ########.fr       */
+/*   Updated: 2022/03/29 15:41:16 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "exe_cmd.h"
 
 static int	read_sign(char c)
 {
