+++ srcs/executor/exe_cmd/exit_utils.c
@@ -1,28 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   env.c                                              :+:      :+:    :+:   */
+/*   exit_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/25 20:01:02 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/03/25 20:01:07 by ynakashi         ###   ########.fr       */
+/*   Created: 2022/03/25 20:01:30 by ynakashi          #+#    #+#             */
+/*   Updated: 2022/03/29 15:41:10 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "exe_cmd.h"
 
-int	my_env(t_envlist *envlist)
+void	numeric_argument_required(char *str)
 {
-	while (envlist)
-	{
-		if (envlist->value == NULL)
-		{
-			envlist = envlist->next;
-			continue ;
-		}
-		printf("%s=%s\n", envlist->key, envlist->value);
-		envlist = envlist->next;
-	}
-	return (0);
+	ft_putendl_fd("exit", STDERR_FILENO);
+	ft_putstr_fd("my_shell: exit: ", STDERR_FILENO);
+	ft_putstr_fd(str, STDERR_FILENO);
+	ft_putendl_fd(": numeric argument required", STDERR_FILENO);
 }
