+++ srcs/executor/exe_cmd/export.c
@@ -6,11 +6,11 @@
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 20:01:52 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/03/25 20:01:57 by ynakashi         ###   ########.fr       */
+/*   Updated: 2022/03/29 15:41:21 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "exe_cmd.h"
 
 static size_t	list_cnt(t_envlist *envlist)
 {
