+++ srcs/executor/exe_cmd/utils.c
@@ -6,11 +6,11 @@
 /*   By: ynakashi <ynakashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 20:02:25 by ynakashi          #+#    #+#             */
-/*   Updated: 2022/03/25 20:02:30 by ynakashi         ###   ########.fr       */
+/*   Updated: 2022/03/29 15:41:47 by ynakashi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "exe_cmd.h"
 
 void	safe_free(char **ptr)
 {
